+++ parser/create_table_statement.go
@@ -1,0 +1,13 @@
+package parser
+
+import "github.com/VorobevPavel-dev/congenial-disco/tokenizer"
+
+type ColumnDefinition struct {
+	Name     tokenizer.Token
+	Datatype tokenizer.Token
+}
+
+type CreateTableStatement struct {
+	Name tokenizer.Token
+	Cols *[]*ColumnDefinition
+}
